feat: redirect Rambo output to Logger by default
Previously, StepRunner used the Rambo behavior, which streams the
command output to IO.

Removed all references to this behavior (called `log`) from
PipelineRunner, so StepRunner's new default behavior is used when
running steps inside of a pipeline.

Removed two assertions in ExIntegrateTest that expected to see IO output
(this is not ultimately the behavior we expect anyway).

Streaming the output to Logger may prove useful when displaying realtime
results to users in a future UI. If not, we can stream it somewhere else
of our choosing.

Fix syntax error in `ExIntegrate.Core.Pipeline.t()`.